rhdl_fpga::axi4lite::core::controller

Module blocking

AXI4Lite Blocking ReadWrite Controller

Purpose

This core provides a simple blocking interface to issue a sequence of read and write requests to an AXI bus using a single input and output stream. Each request is blocked until the response is received, so that only one transaction is in flight at a time.

Schematic Symbol

The core sinks a stream of BlockRequests and then sources a stream of BlockResponse. The BlockRequest stream will generally be stalled, as new requests are only accepted when the previous request has completed.

Internal Details

Internally, the BlockReadWriteController splits the request streams based on the variant of the BlockRequest but will also blank/void out the request if it is waiting for the previous transaction to complete. The backpressure signal for the output (response) stream comes from the external resp_ready signal.

Structs

BlockReadWriteController

Blocking AXI Read/Write Controller

In

Input for the BlockReadWriteController

Out

Output for the BlockReadWriteController

Enums

BlockRequest

Make a blocking read or write request on an AXI bus

BlockResponse

The response to a BlockRequest