rhdl_fpga::axi4lite::core

Module endpoint

Cores used to implement AXI4Lite endpoints

These cores are used to implement an endpoint for an AXI4L bus. They are helpful for building cores that have AXI interfaces on them. For example, to build a core that supports read operations via AXI:

Or to build a core that supports write operations via AXI.

There are complementary cores used for building controllers available in a seperate module.

Modules

read

AXI4Lite Read Endpoint Interface

write

AXI4Lite Write Endpoint Interface