rhdl_fpga::axi4lite::core::endpoint

Module read

AXI4Lite Read Endpoint Interface

Purpose

This core provides a way to interface RHDL stream cores to the AXI4Lite bus interface. This interface core simply repackages the signals to and from the AXI bus into RHDL streams. The addresses are provided on a source stream, and the result of the read (including error conditions) is fed to a sink stream.

Schematic Symbol

Here is the symbol for the core. It provides a source stream for the addresses (which are b32), and sinks a stream of values (which are ReadResult).

Internal Details

Internally, the [Axi2ReadStreams] consists of the following rough contents. A crate::stream::map::Map core is used to map the ReadResult to a ReadResponse, which is then adapted out to the AXI bus.

The Axi2Rhdl and Rhdl2Axi cores buffer their inputs (and outputs) so as to be spec compliant (i.e., no combinatorial logic on the bus is allowed in AXI).

Example

An example of using a [ReadController] and ReadEndpoint together in a test harness is included here:

Non-synthesizable functions are used to generate the request addresses and the replies for demonstration purposes.

//! Test harness for Read Controller and Read Endpoint
//!

use std::iter::repeat_n;

use rhdl::prelude::*;
use rhdl_fpga::{
    axi4lite::{
        core::{controller::read::ReadController, endpoint::read::ReadEndpoint},
        types::ReadResult,
    },
    doc::write_svg_as_markdown,
    rng::xorshift::XorShift128,
    stream::testing::{sink_from_fn::SinkFromFn, source_from_fn::SourceFromFn, utils::stalling},
};

#[derive(Clone, Synchronous, SynchronousDQ)]
struct TestFixture {
    req_source: SourceFromFn<b32>,
    controller: ReadController,
    endpoint: ReadEndpoint,
    req_sink: SinkFromFn<b32>,
    reply_source: SourceFromFn<ReadResult>,
    reply_sink: SinkFromFn<ReadResult>,
}

impl SynchronousIO for TestFixture {
    type I = ();
    type O = ();
    type Kernel = kernel;
}

#[kernel]
pub fn kernel(_cr: ClockReset, _i: (), q: Q) -> ((), D) {
    let mut d = D::dont_care();
    // Wire the request source to the read controller
    d.controller.req_data = q.req_source;
    d.req_source = q.controller.req_ready;
    // Wire the request sink to the read endpoint
    d.req_sink = q.endpoint.req_data;
    d.endpoint.req_ready = q.req_sink;
    // Wire the controller to the reply sink
    d.reply_sink = q.controller.resp_data;
    d.controller.resp_ready = q.reply_sink;
    // Wire the endpoint to the reply source
    d.endpoint.resp_data = q.reply_source;
    d.reply_source = q.endpoint.resp_ready;
    // Wire the AXI busses together.
    d.controller.axi = q.endpoint.axi;
    d.endpoint.axi = q.controller.axi;
    ((), d)
}

fn main() -> Result<(), RHDLError> {
    let rng = XorShift128::default().map(|x| bits(x as u128));
    let address_sink = rng.clone();
    let address = stalling(rng.clone(), 0.23);
    let reply = rng.clone().map(ReadResult::Ok);
    let reply_sink = reply.clone();
    let reply = stalling(reply, 0.23);
    let uut = TestFixture {
        req_source: SourceFromFn::new(address),
        controller: ReadController::default(),
        endpoint: ReadEndpoint::default(),
        req_sink: SinkFromFn::new_from_iter(address_sink, 0.3),
        reply_source: SourceFromFn::new(reply),
        reply_sink: SinkFromFn::new_from_iter(reply_sink, 0.1),
    };
    let input = repeat_n((), 250);
    let input = input
        .with_reset(1)
        .clock_pos_edge(100)
        .take_while(|t| t.time < 1500);

    let vcd = uut.run_without_synthesis(input)?.collect::<Vcd>();
    let options = SvgOptions::default().with_filter(".*controller.*axi.*");
    write_svg_as_markdown(vcd, "axi_read.md", options)?;
    Ok(())
}

with a trace file

Structs

In

Input for the [AxiReadStreams] core

Out

Output from the [AxiReadStreams] core

ReadEndpoint

AXI Read Endpoint