rhdl_fpga::axi4lite::core::switch

Module read

AXI4Lite Read Switch

Purpose

This core provides a way to connect multiple AXI Read endpoints to a single Read manager. Each AXI Read endpoint has an address range that it receives. The address decode logic is user provided in the form of a pure (synthesizable) function.

Internals

The switch is fairly complicated, so the internals are described in pieces. The first part is the ingest pipeline, which maps the incoming requests into decoded (port,addr) values, and handles the case that the decode logic indicates there is an error in the request. The egress pipeline is unbuffered, but includes a combinatorial circuit to detect when responses have been accepted by the endpoint.

Structs

In

Input for the Read switch

Out

Output from the Read Switch

ReadSwitch

AXI Read Switch

Enums

State

The state of the switch

Type Aliases

Command

The input address along with the port to send it to