rhdl_fpga::axi4lite::register

Module rom

A Read Only Register bank on the AXI bus

This simple core provides a readable set of register values to an AXI bus. The read values are also exported via a wire bus for use by other (non-AXI) cores.

Schematic Symbol

Here is the symbol for the core. For simplicity, we assume that 8 bits are enough to select one of the registers in the bank).

Internal Details

Internally, the register contains a ReadEndpoint for communication with the AXI bus. The address is checked against a provided value.

Structs

AxiRom

AXI ROM

In

Input for the [AxiRegBank]

Out

Output for the [AxiRegBank]