rhdl_fpga::axi4lite::register

Module single

A Single register on the AXI Bus

A simple dual ported register with an AXI4Lite bus interface. The address of the register is provided when the core is instantiated.

Schematic Symbol

Here is the symbol for the core.

Internal Details

Internally, the register contains both a ReadEndpoint and WriteEndpoint for communication with the AXI bus. The address is checked against a provided value. Write contention (i.e., writes to the register from both the client side and the AXI side on the same clock cycle) are resolved in factor of the client.

Structs

AxiRegister

AXI Register

In

Input for the AxiRegister

Out

Output for the AxiRegister