rhdl_fpga::axi4lite

Module types

The AXI bus types

This module includes basic types and type definitions required by the AXI specification. Although there are clearly more ergonomic ways to represent these values in rhdl (the ResponseKind, for example, could easily be an enum), we use basic raw types here since we need to export these signals to other cores that are likely not implemented in rhdl.

Modules

response_codes

This module response_codes includes definitions for the 4 responses defined in the specification

Structs

MISO

All MISO signals

MOSI

All MOSI signals

ReadMISO

MISO signals for the read interface

ReadMOSI

MOSI signals for the read interface

ReadResponse

The response to a read transaction

StrobedData

A data word to write, along with a strobe for the bytes

WriteCommand

The write command

WriteMISO

MISO signals for the write interface

WriteMOSI

MOSI signals for the write interface

Enums

AXI4Error

An AXI4-Error Enum meant to capture the two cases of SLVERR and DECERR. The AXI4Error enum is easier to work with in rhdl.

ExFlag

Flag to indicate if the operation was exclusive or not

Functions

read_response_to_result

Helper function to recode a ReadResponse into a Result.

result_to_read_response

Helper function to convert a Result into a ReadResponse

result_to_write_response

Convert from a Result to a ResponseKind for write transactions

strobe_to_mask

Convert a strobe into a mask

write_response_to_result

Convert a ResponseKind to a Result for a write transaction

Type Aliases

AxilAddr

AXI Address type

AxilData

AXI Data type

AxilStrobe

AXI Strobe Mask

ReadResult

The result of a Read operation on the AXI bus

ResponseKind

The ResponseKind called out in the AXI specification

WriteResult

The result of a Write operation on the AXI bus