rhdl_fpga::cdc

Module cross_counter

A Clock domain crossing binary counter

Clock domain crossing counter

Purpose

This core provides a counter where the input pulses come from one clock domain, and the output count is in a different clock domain. The count in the output clock domain is guaranteed to lag behind an equivalent count in the input clock domain.

Connections

The W domain is used for the “writer” to the counter, where the counter increments are provided, and the R domain is used for the “reader” of the counter, where the count is read.

The use is straight forward. In the domain where the pulses to be counted originate, you need to provide a clock and reset signal. You also need to provide a clock and reset in the domain where the count is provided.

Internals

This core uses a vector of 1-bit synchronizers, but with a Gray-coded counter to cross the clock domains.
This is safe because the first stage of registers in the synchronizers will sample the Gray-coded signal essentially simultaneously. The Gray-coded signal is guaranteed to have at most one bit changing at any time point. Thus, all bits will be correct when sampled with the possible exception of the bit that is changing at that time. This bit may resolve to the correct value, or it may not. If it does not, the transition will be missed and the counter will be off by one. However, at the next sample point, this bit will be correct. As the counter is monotonic, it will always lag behind the actual count.

Here is a rough diagram of the contents of the block.

The counter is synchronous the input domain. The output count is fed into a Gray coder, which ensures only one bit changes at a time for each input count. The individual bits of the Gray coder can then be passed through 1-bit synchronizers, as only one bit changes at a time. The output is then combinatorially decoded into a count. Note that the output contains combinatorial delays to the output DFFs of the CDCs. A pipeline stage may be needed to isolate that logic if high speed is required.

Example

use rand::random;
use rhdl::prelude::*;
use rhdl_fpga::{
    cdc::cross_counter::{CrossCounter, In},
    doc::write_svg_as_markdown,
};

fn main() -> Result<(), RHDLError> {
    // Start with a stream of pulses
    let red = (0..).map(|_| random::<bool>()).take(100);
    // Clock them on the red domain
    let red = red.with_reset(1).clock_pos_edge(100);
    // Create an empty stream on the blue domain
    let blue = std::iter::repeat(())
        .with_reset(1)
        .clock_pos_edge(79);
    // Merge them
    let inputs = merge(red, blue, |r: (ClockReset, bool), b: (ClockReset, ())| In {
        incr: signal(r.1),
        incr_cr: signal(r.0),
        cr: signal(b.0),
    });
    // Next we create an instance of the clock-domain crossing core, with
    // the appropriate clock domains.
    let uut = CrossCounter::<Red, Blue, 4>::default();
    // Simulate the crosser, and collect into a VCD
    let vcd = uut
        .run(inputs)?
        .take_while(|x| x.time < 1000)
        .collect::<Vcd>();
    let options = SvgOptions {
        label_width: 20,
        ..Default::default()
    };
    write_svg_as_markdown(vcd, "cross_counter.md", options)?;
    Ok(())
}

Structs

CrossCounter

Unit to instantiate.

In

Inputs to the core

Out

Outputs from the core

Functions

cross_counter_kernel

The kernel function for the cross-counter.