rhdl_fpga::cdc

Module synchronizer

A one-bit synchronizer

Single bit synchronizer

Purpose

This core allows you to safely (relatively) cross clock domains with a single bit signal. The problem with crossing clock domains is related to the potential violation of the invariants of flip flops. Consider the following super trivial example of an input to a flip flop that transitions close to the clock edge, within some time δt

The `??` states represent the fact that we cannot know for sure what state the flip flop will be in during this time interval. And worse, any outputs connected to this FF will also be in an indeterminate state.
This phenomenon is known as metastability, and there are plenty of resources on the web available to learn more. The “fix” is to note that for a real flop, it will (quickly) decide one way or the other as to what it’s going to take a true or false value. In that case, we can actually draw the diagram something more like this.

In the upper `q1` trace, the ff decided that the input was `false`, and so went low for the cycle. In the lower `q1` trace, the ff decided that the input was `true` and so went high for the cycle.
This should probably terrify you. It basically means that you get a random bit output because of the violation of the timing for the flip flop. The fix is to sample the output of the ff with another flip flop, (q2), which will at least resolve to one of the two states at every clock edge. The important part here is not that the signal made it out through the second flop. The important part is that there are no ?? sections in the output q2. So even though the bit may be lost, at least the uncertainty stops propagating.

I’ve seen hardware designers put longer chains in place in the belief that this somehow avoids the problem. It does not. The damage was done at the first FF. The subsequent ones are only to reduce the probability of a ?? appearing at the output. This is already exceptionally unlikely as the amount of time the FF spends in the ?? state is generally quite short with respect to the clock cycle period. However, a belt-and-suspenders approach suggests an extra FF doesn’t hurt.

The root problem, however, remains. The bit was lost in transition. So what to do? Simplest answer is Do Not Use This. It’s best thought of as an unsafe primitive out of which safe cores can be built. You need to encode the signal crossing the clock boundary so that effectively, this phenom doesn’t happen. If you want to cross a counter (so that the output clock has the count of pulses arriving at the input clock domain), then look at CrossCounter. If you want to send data between clock domains, use a FIFO, such as AsyncFIFO.

Finally, if you only want to cross data relatively slowly, use a multi-bit handshake based method, like the [SlowCrosser].

For more detail, see [this doc]! It has lots of great detail. (http://cva.stanford.edu/people/davidbbs/classes/ee108a/winter0607%20labs/lect.9.Metastability-blackschaffer.ppt)

And just in case you think I am exaggerating the danger here (which is very analogous to a memory safety issue), here we end up with a pulse that may or may not arrive in the new clock domain at all!

Finally: Note that rhdl does not simulate metastability. Essentially for DFFs in rhdl, the value of δt is 0. So be extra cautious when assuming that your designs are safe if you are using synchronizers as raw components.

Connections

Here is the schematic symbol for the synchronizer.

Internals

Internally, the structure is simple:

Example

Here is a simple example of a 1 bit synchronizer being fed random pulses

use rand::random;
use rhdl::prelude::*;
use rhdl_fpga::{
    cdc::synchronizer::{In, Sync1Bit},
    doc::write_svg_as_markdown,
};

fn main() -> Result<(), RHDLError> {
    // Start with a stream of pulses
    let red = (0..).map(|_| random::<u8>() > 200).take(100);
    // Clock them on the red domain
    let red = red.with_reset(1).clock_pos_edge(100);
    // Create an empty stream on the blue domain
    let blue = std::iter::repeat(())
        .with_reset(1)
        .clock_pos_edge(79);
    // Merge them
    let inputs = merge(red, blue, |r: (ClockReset, bool), b: (ClockReset, ())| In {
        data: signal(r.1),
        cr: signal(b.0),
    });
    // Next we create an instance of the 1-bit synchronizercore, with
    // the appropriate clock domains.
    let uut = Sync1Bit::<Red, Blue>::default();
    // Simulate the crosser, and collect into a VCD
    let vcd = uut
        .run(inputs)?
        .take_while(|x| x.time < 2000)
        .collect::<Vcd>();
    let options = SvgOptions {
        label_width: 20,
        ..Default::default()
    };
    write_svg_as_markdown(vcd, "sync_cross.md", options)?;
    Ok(())
}

With an output trace

Structs

In

Input to the synchronizer

Sync1Bit

A simple two-register synchronizer for crossing a single bit from the W domain to the R domain