rhdl_fpga::core

Module constant

Constant driver

This core provides a constant value that can be provided later in the compilation phase (i.e., not when rustc runs).

The schematic symbol is simple:

There is no timing information, the constant core simply provides the constant value all the time.

Example

Here is an example of the constant being used.

use rhdl::prelude::*;
use rhdl_fpga::{core::constant::Constant, doc::write_svg_as_markdown};

fn main() -> Result<(), RHDLError> {
    let inputs = std::iter::repeat(()).without_reset().clock_pos_edge(100);
    let uut = Constant::new(b8(42));
    let vcd = uut
        .run(inputs)?
        .take_while(|x| x.time < 500)
        .collect::<Vcd>();
    let options = SvgOptions {
        label_width: 20,
        ..Default::default()
    };
    write_svg_as_markdown(vcd, "constant.md", options)?;
    Ok(())
}

The simulation trace is pretty boring.

Structs

Constant

The core to include for the constant driver