rhdl_fpga::core

Module delay

Delay Line

This module implements a delay line in which the data type propagating through the delay is generic of type T, and the length of the delay is compile time configurable.

Here is the schematic symbol

# Internals Internally the delay is simply a chain of `N` [DFF]s, in a linear chain. Note that the flip flops will reset to the default value for `T`, which is why it is required for `T: Default`.

Example

The delay is a fairly simple core, and the example is pretty basic. To make it slightly more interesting, we demonstrate the case that the data being carried is an enum.

use rhdl::prelude::*;
use rhdl_fpga::{core::delay::Delay, doc::write_svg_as_markdown};

#[derive(PartialEq, Digital, Default)]
pub enum State {
    Init,
    Running,
    Busted,
    Reset,
    #[default]
    Unknown,
}

fn main() -> Result<(), RHDLError> {
    let input = [
        State::Unknown,
        State::Init,
        State::Reset,
        State::Busted,
        State::Running,
        State::Busted,
    ]
    .into_iter()
    .chain(std::iter::repeat_n(State::Unknown, 4));
    let input = input.with_reset(1).clock_pos_edge(100);
    let uut: Delay<State, 3> = Delay::default();
    let vcd = uut.run(input)?.collect::<Vcd>();
    let options = SvgOptions::default()
        .with_filter("(^top.input.*)|(^top.output.*)")
        .with_label_width(20);
    write_svg_as_markdown(vcd, "delay.md", options)?;
    Ok(())
}

The resulting trace is show below.

Structs

Delay

The Delay core T is the type carried by the core N is the length of the delay line

Functions

delay

Kernel for delay core