rhdl_fpga::core::option

Function pack

pub fn pack<T: Digital>(valid: bool, data: T) -> Option<T>

Packs a tag and a data value back into an Option.
The data value argument is ignored if the tag is false.