rhdl_fpga::core::ram

Module asynchronous

An asynchornous block RAM

A simple block ram that stores 2^N values of type T. It has two interfaces for read and writing, and supports two different clocks. This RAM is meant primarily for FPGAs, as you can specify the initial contents of the RAM. For ASICs, you should probably assume the contents of the RAM are random on initialization and implement reset mechanism.

This block ram is not “combinatorial” but is rather “fully registered”. That means that the read address is sampled on the positive edge of the read clock, and the data is presented prior to the positive edge of the next clock.

There are block rams that don’t have this limitation (i.e., they provide the read output on the same clock as the read address). But they are generally not portable. If you need one of those, you should create a custom model for it.

Here is the schematic symbol:

Timing

From a timing perspective, the asynchronous RAM operates with two clocks, a read clock and a write clock. The output is synchronous with the read clock, as indicated by the type signature. The write inputs are synchronous to their own write clock.

It is considered undefined behavior to read and write simultaneously to the same address in the block RAM. With multiple clocks, it becomes even more nebulous to define what “simultaneous” means. In general, you should avoid accessing the same address/row of the RAM from both the read and write sides of the interface at the same time.

Here is a simple timing diagram:

In general, I don’t recommend using an AsyncBRAM. It’s easier and more idiomatic to use an [OptionAsyncBRAM] or a [PipeAsyncBRAM] which provide Option-based interfaces.

Example

Here is an example running the timing pattern shown above.

use rhdl::prelude::*;
use rhdl_fpga::{
    core::ram::asynchronous::{AsyncBRAM, In, ReadI, WriteI},
    doc::write_svg_as_markdown,
};

fn main() -> Result<(), RHDLError> {
    let read = [b3(1), b3(2), b3(1), b3(1)];
    let write = [None, Some((b3(1), 42)), None, None, None, None];
    let read = read.into_iter().without_reset().clock_pos_edge(100).map(|t| {
        t.map(|(cr, val)| ReadI {
            addr: val,
            clock: cr.clock,
        })
    });
    let write = write.into_iter().without_reset().clock_pos_edge(79).map(|t| {
        t.map(|(cr, val)| match val {
            Some((addr, data)) => WriteI {
                addr,
                data: bits(data),
                enable: true,
                clock: cr.clock,
            },
            None => WriteI {
                addr: b3(0),
                data: b8(0),
                enable: false,
                clock: cr.clock,
            },
        })
    });
    // Merge them
    let input = merge(read, write, |r, w| In {
        read: signal(r),
        write: signal(w),
    });
    let uut: AsyncBRAM<b8, Red, Blue, U3> = AsyncBRAM::new((0..).map(|x| (b3(x), b8(x))));
    let vcd = uut.run(input)?.collect::<Vcd>();
    let options = SvgOptions {
        label_width: 20,
        ..Default::default()
    };
    write_svg_as_markdown(vcd, "async_bram.md", options)?;
    Ok(())
}

With a simulation trace:

Structs

AsyncBRAM

The AsyncBRAM core. It stores elements of type T. The write side is on the W clock domain, and the read side is on the R clock domain. The N parameter indicates the number of address bits which determines the size of the BRAM.

In

The inputs for the AsyncBRAM core

ReadI

For the input interface, we have write and read parts.
These are on different clock domains, so we need to split them out. The read input lines contain the current address and the clock signal.

WriteI

The write input lines control the write side of the RAM. It contains the address to write to, the data, and the enable and clock signal.