rhdl_fpga::core::ram

Module option_async

An asynchronous RAM with an Option<T interface

This version of the asynchornous BRAM replaces the write interface with an Option driven interface to make it more idiomatic to use in RHDL. The schematic symbol looks like this

Internally, the circuitry is composed of a combinatorial unpacker.

Example

Here is the simple example that demonstrates how to use the option interface.

use rhdl::prelude::*;
use rhdl_fpga::{
    core::ram::{
        asynchronous::ReadI,
        option_async::{In, OptionAsyncBRAM, WriteI},
    },
    doc::write_svg_as_markdown,
};

fn main() -> Result<(), RHDLError> {
    let read = [b3(1), b3(2), b3(1), b3(1)];
    let write = [None, Some((b3(1), b8(42))), None, None, None, None];
    let read = read
        .into_iter()
        .without_reset()
        .clock_pos_edge(100)
        .map(|t| {
            t.map(|(cr, val)| ReadI {
                addr: val,
                clock: cr.clock,
            })
        });
    let write = write
        .into_iter()
        .without_reset()
        .clock_pos_edge(79)
        .map(|t| {
            t.map(|(cr, val)| WriteI {
                data: val,
                clock: cr.clock,
            })
        });
    // Merge them
    let input = merge(read, write, |r, w| In {
        read: signal(r),
        write: signal(w),
    });
    let uut: OptionAsyncBRAM<b8, Red, Blue, U3> =
        OptionAsyncBRAM::new((0..).map(|x| (b3(x), b8(x))));
    let vcd = uut.run(input)?.collect::<Vcd>();
    let options = SvgOptions {
        label_width: 20,
        ..Default::default()
    };
    write_svg_as_markdown(vcd, "option_async_bram.md", options)?;
    Ok(())
}

With a resulting trace file here.

Structs

In

The input struct for the OptionAsyncBRAM

OptionAsyncBRAM

The unit to include for an option interface to the asynchronous BRAM.

WriteI

The write interface for the OptionAsyncBRAM.

Functions

ram_kernel

Kernel function for OptionAsyncBRAM