rhdl_fpga::core::ram

Module option_sync

A synchronous ram wrapped in an interface that accepts Option<T> for writing

This version of the BRAM replaces the write enable with an Option<T> interface which is more idiomatic for RHDL. THe schematic symbol looks like this:

Internally, the circuitry is very simple. The option is decoded (combinatorially) into a write enable flag, an address and a data value. These are then forwarded to the regular synchronous BRAM.

# Example
Here is a simple example that issues reads and writes using the two interfaces. The setup for the read is

where the last cycle is included to allow for the latency of the read in cycle 4.

use rhdl::prelude::*;
use rhdl_fpga::{
    core::ram::option_sync::{In, OptionSyncBRAM},
    doc::write_svg_as_markdown,
};

fn main() -> Result<(), RHDLError> {
    let read = [b4(4), b4(5), b4(2), b4(1), b4(4), b4(4)];
    let write = [Some((b4(2), b8(42))), None, Some((b4(4), b8(21)))]
        .into_iter()
        .chain(std::iter::repeat(None));
    let inputs = read.into_iter().zip(write).map(|(r, w)| In {
        read_addr: r,
        write: w,
    });
    let inputs = inputs.with_reset(1).clock_pos_edge(100);
    let uut = OptionSyncBRAM::<b8, U4>::new((0..).map(|ndx| (b4(ndx), b8(ndx))));
    let vcd = uut.run(inputs)?.collect::<Vcd>();
    let options = SvgOptions {
        label_width: 20,
        ..Default::default()
    };
    write_svg_as_markdown(vcd, "option_bram.md", options)?;
    Ok(())
}

With a resulting trace file here.

Structs

In

The input struct for the OptionSyncBRAM

OptionSyncBRAM

The unit to include that wraps the [SyncBRAM] The T parameter indicates the type of element stored in the BRAM. It must implement Digital. The N parameter indicates the number of address bits. Thus, the BRAM will hold 2^N elements.

Functions

ram_kernel

Kernel function for OptionSyncBRAM