rhdl_fpga::core::ram

Module pipe_sync

A pipeline interface to a BRAM

This is a pipeline RAM interface. It provides the ability to read and write from the RAM using the Option<T> interface for a non-stallable pipeline. What this means is that you can think of reading from the RAM as injecting a command to read from a given address A, and then some time later, out comes the corresponding data element D.

The schematic symbol for this unit looks like this:

where `?` is short for `Option`.
The delay between the injection of A and the receipt of D is fixed (one clock cycle for the BRAM). It could also be variable (as in accessing a DRAM, for example). But in this case it is fixed and constant. This circuit tracks the single cycle latency on the input and uses it to blank out the output. Like this:

From this diagram, it is clear what the internal guts of the pipeline RAM look like. The tag (valid bit) is stripped off of the input signal, delayed by one clock cycle to account for the latency of the BRAM read, and then merged with the output value.

Example

Here is an example of the the PipeSyncBRAM being used with the same test sequence illustrated above.

use rhdl::prelude::*;
use rhdl_fpga::core::ram::pipe_sync::{In, PipeSyncBRAM};

fn main() -> Result<(), RHDLError> {
    // Generate the stream example from the timing diagram.
    // Read location 2, then 3 then 2 again
    let reads = [None, Some(b3(2)), Some(b3(3)), Some(b3(2)), None];
    // Write to location 2 while reading from 3
    let writes = [None, None, Some((b3(2), b8(42))), None, None];
    let inputs = reads
        .into_iter()
        .zip(writes)
        .map(|(r, w)| In { read: r, write: w })
        .with_reset(1)
        .clock_pos_edge(100);
    let uut = PipeSyncBRAM::new((0..).map(|x| (b3(x), b8(x))));
    let vcd = uut.run(inputs)?.collect::<Vcd>();
    rhdl_fpga::doc::write_svg_as_markdown(
        vcd,
        "pipe_ram.md",
        SvgOptions::default()
            .with_label_width(20)
            .with_filter("(^top.clock.*)|(^top.input.*)|(^top.output.*)"),
    )
    .unwrap();
    Ok(())
}

The resulting trace file is:

Structs

In

The inputs for the [PipeBRAM]

PipeSyncBRAM

The unit that implements the [PipeBRAM] The T parameter indicates the type of data held in the BRAM. The N parameter indicates the number of address bits.

Functions

kernel

The kernel for the [PipeBRAM]