rhdl_fpga::core::ram

Module synchronous

A simple synchronous block ram.

The contents are generic over a type T, and the address is assumed to be N bits wide.

Here is the schematic symbol

# Timing From a timing perspective, it is assumed that the block ram implements a single cycle delay on both the read and write interfaces. Reading and writing to the same address cell at the same clock cycle does not result in defined behavior. The underlying primitive may provide some guarantees (like write before read), but these are not enforced, so use with caution.
The following diagram illustrates the basic timing. The Data@A1 indicates the nominal contents of the BRAM cell at address A1:

In general, I don’t recommend using a SyncBRAM. It’s easier and more idiomatic to use either OptionSyncBRAM or PipeSyncBRAM, which provide Option interfaces.

Example

Here we encode the example above.

use rhdl::prelude::*;
use rhdl_fpga::{
   core::ram::synchronous::{In, SyncBRAM, Write},
   doc::write_svg_as_markdown,
};

fn main() -> Result<(), RHDLError> {
   let read = [b4(1), b4(2), b4(1), b4(1)];
   let no_write = Write {
       addr: b4(0),
       value: b8(0),
       enable: false,
   };
   let write = [
       no_write,
       Write {
           addr: b4(1),
           value: b8(42),
           enable: true,
       },
       no_write,
       no_write,
   ];
   let input = read
       .into_iter()
       .zip(write)
       .map(|(r, w)| In {
           read_addr: r,
           write: w,
       })
       .with_reset(1)
       .clock_pos_edge(100);
   let uut = SyncBRAM::new((0..).map(|x| (b4(x), b8(x))));
   let vcd = uut.run(input)?.collect::<Vcd>();
   let options = SvgOptions::default();
   write_svg_as_markdown(vcd, "sync_bram.md", options)?;
   Ok(())
}

The trace below demonstrates the result.

Structs

In

Core inputs

SyncBRAM

The synchronous version of the block ram.

Write

A collection of signals for a raw write interface