rhdl_fpga::dsp::lerp

Module fixed

Linear Interpolation Functions

In DSP work, you often need to linearly interpolate between two values. These functions provide that computation, with a variable number of bits in the two input arguments and a variable number of bits in the interpolation factor.

Internal Details

There are various subtleties at play here.
Given A: Bits, B: Bits and a factor: Bits,

we want to compute

A * (1 - delta) + B * delta = Y

where

0 <= delta = factor / 2^M < 1

The < 1 part is important, since with M bits, it is not possible to represent 2^M. The largest value that delta can take is 2^(M-1)/2^M. Normally, when linear interpolation is used, this limitation is not a problem. However, you should be aware, in case you need the function to be able to handle the case of delta = 1.

Substituting delta, we get

A * ( 1 - factor / 2^M) + B * factor / 2^M = Y

Multiplying out by 2^M, we get

A * 2^M - A * factor + B * factor = Y * 2^M

To get this into a single multiplication, we need

A * 2^M + (B - A) * factor = Y * 2^M

Even if B and A are unsigned, the B - A term is signed, so we need to promote the factor to be signed as well

A * 2^M + Diff * signed_factor = Y * 2^M

Here signed_factor`` will be M+1`` bits wide, and Diff will be N+1 bits wide

The product will thus be M+N+2 bits wide. The factor A * 2^M will be N+M bits wide, and is unsigned. So we need to convert it to a signed value (which adds 1 bit) and then extend it (signed) by a bit.

We can (after adding it), right shift by M bits to retrieve Y, and then truncate the value to N bits, and safely cast as unsigned. Because the output is guaranteed to be unsigned and the number of bits cannot increase (ignoring fractional bits, of course), this operation can be safely carried out with a truncation operation.

Example

The following example demos the unsigned lerp function, by wrapping it into a core using the Func wrapper.

use rhdl::prelude::*;
use rhdl_fpga::doc::write_svg_as_markdown;
use rhdl_fpga::dsp::lerp::fixed::lerp_unsigned;

// Because the `lerp` function is just a function, if we
// want to make a core out of it, we need a wrapper. In
// this example, we will use the [Func] wrapper.  We
// need to fit the type signature of that function, so the
// inputs of the `lerp` function need to be put into a single
// struct.
#[derive(PartialEq, Digital)]
pub struct LerpIn<N, M>
where
    N: BitWidth,
    M: BitWidth,
{
    pub lower_value: Bits<N>,
    pub upper_value: Bits<N>,
    pub factor: Bits<M>,
}

// An wrapper function to call the `lerp_unsigned`
#[kernel]
pub fn wrap_lerp<N, M>(_cr: ClockReset, i: LerpIn<N, M>) -> Bits<N>
where
    N: BitWidth,
    M: BitWidth,
{
    lerp_unsigned::<N, M>(i.lower_value, i.upper_value, i.factor)
}

fn main() -> Result<(), RHDLError> {
    // The [Func] wrapper gives us a core we can simulate
    let uut: Func<LerpIn<U8, U4>, Bits<U8>> = Func::try_new::<wrap_lerp<U8, U4>>()?;
    // Simulate a ramp
    let ramp = (0..15)
        .map(|x| LerpIn {
            upper_value: bits(255),
            lower_value: bits(0),
            factor: bits(x),
        })
        .without_reset()
        .clock_pos_edge(100);
    let vcd = uut.run(ramp)?.collect::<Vcd>();
    write_svg_as_markdown(vcd, "lerp.md", SvgOptions::default())?;
    Ok(())
}

The resulting trace shows the linear interpolation

Functions

lerp_signed

Linearly interpolate between signed values

lerp_unsigned

Linearly interpolate between unsigned values