rhdl_fpga::fifo

Module synchronous

A simple synchronous FIFO.

This FIFO is designed to be as simple as possible and thus be robust. It is a two-port FIFO, with separate read and write ports. The FIFO is parameterized by the number of bits in each element. The depth of the FIFO is 2^N-1 elements. You cannot fill the FIFO to 2^N elements.

Here is the schematic symbol for the FIFO

Example

Testing a synchronous FIFO is a little tricky, since the input and output interfaces have feedback, and thus do not lend themselves to the regular input -> uut -> output type of testing.

Instead, we can either build custom testing harnesses (see [SyncFIFOTester]) or use the feedback testing mechanism. The feedback testing mechanism allows you to provide a closure that computes the next set of inputs given the current outputs. For synchronous circuits, like this one, the clock is handled for you, so your function need not worry about clock edges. Here is an example.

use rhdl::{core::sim::ResetOrData, prelude::*};
use rhdl_fpga::{doc::write_svg_as_markdown, fifo::synchronous::SyncFIFO};

fn main() -> Result<(), RHDLError> {
    type Uut = SyncFIFO<Bits<U16>, U4>;
    let uut = Uut::default();
    let mut need_reset = true;
    let test_seq = (0..)
        .map(|_| b16(rand::random::<u16>() as u128))
        .take(100)
        .collect::<Vec<_>>();
    let mut input_seq = test_seq.iter().copied();
    let mut output_seq = test_seq.iter().copied();
    let vcd = uut
        .run_fn(
            |output| {
                // Handle the reset pulse at the begining
                if need_reset {
                    need_reset = false;
                    return Some(ResetOrData::Reset);
                }
                let mut input = <Uut as SynchronousIO>::I::dont_care();
                // By default, we do not insert more data.
                input.data = None;
                // The FIFO is not full, so we can insert data.  Toss a coin.
                if !output.full && rand::random::<bool>() {
                    input.data = input_seq.next();
                }
                // By default, we do not advance the FIFO
                input.next = false;
                // The FIFO has valid data.  So we can receive data. Toss a coin.
                if output.data.is_some() && rand::random::<bool>() {
                    input.next = true;
                    assert_eq!(output_seq.next().unwrap(), output.data.unwrap());
                }
                // Return the constructed input
                Some(ResetOrData::Data(input))
            },
            100,
        )
        .take_while(|t| t.time < 1500)
        .collect::<Vcd>();
    let options = SvgOptions::default().with_filter("(^top.input(.*))|(^top.outputs(.*))");
    write_svg_as_markdown(vcd, "sync_fifo.md", options)?;
    Ok(())
}

The trace below demonstrates the result.

Structs

In

Inputs for the FIFO

Out

Outputs from the FIFO

SyncFIFO

A simple synchronous FIFO T is the data type held by the FIFO. Note that we need T: Default. N the number bits in the address. FIFO holds 2^{N-1} elements when full.

Functions

fifo_kernel

The compute kernel for the SyncFIFO