rhdl_fpga::fifo::testing

Module filler

Bursty FIFO Filler

A bursty, random FIFO filler. Uses a sequence of values from an XorShift128 to fill a FIFO. The lowest N bits of the output number are used as the data. Based on the random value, the filler will also decide to “sleep” for a number of clock cycles. This is to simulate a bursty data source. Note that the behavior is deterministic. The number of sleep cycles is also fixed, so that a single parameter can be used to control the “burstiness” of the data.

Schematic Symbol

The FIFOFiller has the following symbol:

Internally, the FIFOFiller uses an XorShift core to generate a sequence of pseudorandom 32 bit values. These are used to both generate the output data, and to determine if the core will sleep.

Structs

FIFOFiller

The FIFO Filler core

In

Inputs to the FIFOFiller core

Out

Outputs from the FIFOFiller core