rhdl_fpga::lid

Module carloni

A Carloni Relay station (aka Skid Buffer)

Implement a Carloni Relay station. As described in the paper “From Latency-Insensitive Design to Communication-Based System-Level Design” by Carloni. Proceedings of the IEEE, 2015.

This is an implementation of the relay station as shown in Figure 4. A relay station is essentially a set of flip flops that can be inserted into a Ready-Valid bus to break long paths. They add latency, but do not affect overall throughput.

Schematic symbol

Here is a symbol for the buffer, as depicted in the original paper.

The `void` signal indicates the validity of the data line, and `stop` is an active-high flow control mechanism.
Internal details

The details of the buffer are well explained in the paper. Here is rough sketch of the internals of the buffer component.

Note that the flip flops in this design have enable signals. The logic of the control block is shown below.

For more details, see the paper.
Example

The following example uses the run_fn method on Synchronous circuits to provide a feedback-type test harness (i.e., the input depends on the outputs of the circuit at the last clock edge).

use rhdl::prelude::*;
use rhdl_fpga::{
    lid::carloni::{self, Carloni},
    rng::xorshift::XorShift128,
};

fn main() -> Result<(), RHDLError> {
    // The buffer will manage items of 4 bits
    let uut = Carloni::<b4>::default();
    // The test harness will include a consumer that
    // randomly pauses the upstream producer.  The producer
    // will also randomly pause it's output.
    let mut need_reset = true;
    let mut source_rng = XorShift128::default();
    let vcd = uut
        .run_fn(
            |out| {
                if need_reset {
                    need_reset = false;
                    return Some(rhdl::core::sim::ResetOrData::Reset);
                }
                let mut input = carloni::In::<b4>::dont_care();
                // See if we want to pause the stream as the consumer
                let want_to_pause = rand::random::<u8>() > 200;
                input.stop_in = want_to_pause;
                // Decide if the producer will generate a data item
                let want_to_send = rand::random::<u8>() < 200;
                input.void_in = true;
                input.data_in = bits(0);
                if !out.stop_out && want_to_send {
                    // The receiver did not tell us to stop, and
                    // we want to send something.
                    input.data_in = bits((source_rng.next().unwrap() & 0xF) as u128);
                    input.void_in = false;
                }
                Some(rhdl::core::sim::ResetOrData::Data(input))
            },
            100,
        )
        .take_while(|t| t.time < 1500)
        .collect::<Vcd>();
    rhdl_fpga::doc::write_svg_as_markdown(vcd, "carloni.md", SvgOptions::default())?;
    Ok(())
}

Here is the trace

Structs

Carloni

A Carloni (skid) buffer

In

The inputs for the Carloni buffer

Out

The outputs for the Carloni buffer