rhdl_fpga::pipe

Module chunked

Chunked Pipe Core

Purppose

A Chunked Pipe Core takes a sequence of T data elements and chunks them into an array of size N. It is equivalent to calling .chunks() on slices. Note that each chunk will contain a disjoint set of samples.

If you want a sliding window, use the [WindowedPipe] Core instead.
Schematic Symbol

Here is the schematic symbol for the [ChunkedPipe] core.

Internals

The Chunked core includes a pipeline delay stage with taps to extract the delayed signals. An input buffer is needed at the input to isolate the delay chain from the input pipeline.

The control system simply counts the number of valid elements in the tapped delay line and sets the output tag to true for one cycle when the count is equal to N.

Example

use rhdl::prelude::*;
use rhdl_fpga::{
    pipe::chunked::Chunked, rng::xorshift::XorShift128, stream::testing::utils::stalling,
};

fn main() -> Result<(), RHDLError> {
    let uut = Chunked::<U2, b4, 4>::default();
    let source_rng = XorShift128::default().map(|x| bits((x & 0xF) as u128));
    let input = stalling(source_rng, 0.1)
        .with_reset(1)
        .clock_pos_edge(100)
        .take_while(|t| t.time < 1500);
    let vcd = uut.run(input)?.collect::<Vcd>();
    rhdl_fpga::doc::write_svg_as_markdown(
        vcd,
        "pipe_chunked.md",
        SvgOptions::default().with_io_filter(),
    )?;
    Ok(())
}

Structs

Chunked

The Chunked Pipe Core

Type Aliases

In

Inputs for the Chunked Pipe

Out

Outputs for the Chunked Pipe