rhdl_fpga::pipe

Module filter

Filter Pipe Core

Purpose

A Filter Pipe core takes a pipeline of elements of type T and a function fn(T) -> bool, and keeps only those items for which the function evaluates to true. The filter funciton is provided in the form of a synthesizable function. This is equivalent to using .filter() on an iterator.

Schematic Symbol

Here is a schematic symbol for the Filter core.

Internals

The Filter core includes a buffer to separate the combinatorial filter function from the previous pipe stage, but not an output buffer. This way, when pipe stages are connected together, the total latency only increases by one per stage added.

Example

Here is an example of a filtering pipeline, in which some of the elements are discarded by the filter function. The single cycle of latency introduced by the input buffer is used to isolate the combinatorial pathways.

use rhdl::prelude::*;
use rhdl_fpga::{
    pipe::filter::Filter, rng::xorshift::XorShift128, stream::testing::utils::stalling,
};

#[kernel]
fn keep_even(_cr: ClockReset, t: b4) -> bool {
    !(t & bits(1)).any()
}

fn main() -> Result<(), RHDLError> {
    let input = XorShift128::default().map(|x| b4((x & 0xF) as u128));
    let input = stalling(input, 0.23);
    let uut = Filter::try_new::<keep_even>()?;
    let input = input
        .with_reset(1)
        .clock_pos_edge(100)
        .take_while(|t| t.time < 1500);
    let vcd = uut.run(input)?.collect::<Vcd>();
    rhdl_fpga::doc::write_svg_as_markdown(vcd, "pipe_filter.md", SvgOptions::default())?;
    Ok(())
}

with the trace file:

Structs

Filter

The Filter Pipe Core

Type Aliases

In

The input for the Filter

Out

The output for the Filter