rhdl_fpga::pipe

Module map

Map Pipe Core

Purpose

A Map pipe core takes a pipeline of elements of type T and a synthesizable function fn(T) -> S and outputs a pipeline that carries elements of type S. Pipes do not accept backpressure, but they do accept a strobe in the form of an Option input. This is equivalent to calling .map() on an iterator. Note that the inputs are registered, but not the outputs. This is by design, so that the pipe cores can be chained without redundant register slices between them. If you want an extra register slice, you can insert a [Buffer] in the pipeline. There are no combinatorial paths from the input to the output, and this is verified via a test.

Schematic Symbol

Here is the schematic symbol for the Map core.

Internals

The Map core is simple. A DFF registers the incoming data. When the input is Some, the provided function is provided the data, and the output wrapped into an option for passing down the pipeline.

Example

Here is an example of a mapping a pipeline, transforming elements. Note that there is a single clock cycle of latency introduced by the input buffer.

use rhdl::prelude::*;
use rhdl_fpga::{
    core::slice::lsbs, pipe::map::Map, rng::xorshift::XorShift128, stream::testing::utils::stalling,
};

#[kernel]
fn map_item(_cr: ClockReset, t: b4) -> b2 {
    lsbs::<U2, U4>(t)
}

fn main() -> Result<(), RHDLError> {
    let input = XorShift128::default().map(|x| b4(x as u128 & 0xF));
    let uut = Map::try_new::<map_item>()?;
    let input = stalling(input, 0.23);
    let input = input
        .with_reset(1)
        .clock_pos_edge(100)
        .take_while(|t| t.time < 1500);
    let vcd = uut.run(input)?.collect::<Vcd>();
    rhdl_fpga::doc::write_svg_as_markdown(vcd, "pipe_map.md", SvgOptions::default())?;
    Ok(())
}

with a trace file like this:

Structs

Map

The Map Core (Pipe Version)

Type Aliases

In

The input for the Map pipe

Out

The output for the Map pipe