rhdl_fpga::reset

Module conditioner

Reset Conditioner

Given a reset signal in domain W that is asynchronous to the clock of domain R, generate a reset signal in domain R that is synchronous to the clock of domain R.

Here is the schematic symbol for the reset conditioner

Internal circuitry

The internal circuitry of the reset conditioner is show below. The reset conditioner is essentially a [Sync1Bit] one-bit synchronizer, with a constant input (and inverted output). The difference is that the resets of both flip flops are tied to the input reset.

Timing

The behavior of the ResetConditioner is roughly depicted below.

The behavior is as follows.

In steady state, the output is zero (due to the inverter) at t0

When the reset signal is asserted, both flops immediately go into reset and the output goes high (again, due to the inverter) at t1

When the reset is released, the circuit behaves like a 1-bit synchronizer that “sees” a transition from 0 -> 1 at that instant at t2.

Depending on the timing of that transition relative to the clock cycle, it may take either 1 or 2 clocks for the transition to propagate to the output. This depends on the difference between t2 and the next positive edge. In this case, we assume it is missed (arrives too close to the edge)

The start and duration of the reset pulse are thus not deterministic.

The end of the reset pulse, however, is aligned to a clock edge, which is (generally) the important characteristic for an synchronous reset. This is the time t3

Example

On it’s own, a ResetConditioner is not particularly useful.
But here is an example of how it operates with resets generated in a different time domain from the one they are being crossed to.

use rand::{Rng, SeedableRng};
use rhdl::prelude::*;
use rhdl_fpga::{
    doc::write_svg_as_markdown,
    reset::conditioner::{In, ResetConditioner},
};

fn sync_stream() -> impl Iterator<Item = TimedSample<In<Red, Blue>>> {
    let mut rng = rand::rngs::StdRng::seed_from_u64(0xdead_beef);
    // Assume the red stuff comes on the edges of a clock
    let red = (0..)
        .map(move |_| rng.random::<u8>() > 150)
        .take(100)
        .with_reset(1)
        .clock_pos_edge(50);
    let blue = std::iter::repeat(false).with_reset(1).clock_pos_edge(79);
    red.merge(blue, |r, g| In {
        reset: signal(reset(r.1)),
        clock: signal(g.0.clock),
    })
}

fn main() -> Result<(), RHDLError> {
    let uut = ResetConditioner::<Red, Blue>::default();
    let input = sync_stream();
    let vcd = uut
        .run(input)?
        .take_while(|t| t.time < 1400)
        .collect::<Vcd>();
    write_svg_as_markdown(vcd, "reset_conditioner.md", SvgOptions::default())?;
    Ok(())
}

The resulting trace file is here.

Structs

In

The inputs to the ResetConditioner.

ResetConditioner

The ResetConditioner circuit. Here W is the domain where the reset circuit originates, and R is the domain where the reset is being sent to (i.e., is synchronous with).