rhdl_fpga::reset

Module negation

Reset Negation Core

Because the Reset type is meant to signal an active-high reset, if you need a reset that is active-low, you will need a ResetNegation core. This core is simply a not gate, but it can unwrap the ResetN type and create a new Reset type (which signifies) an active-high reset. This is sometimes needed for interfacing to external systems that provide a active-low reset.

Schematic

Here is the schematic symbol for the ResetNegation core.

Example

Here is a simple example of using the ResetNegation core.

use rhdl::prelude::*;
use rhdl_fpga::{doc::write_svg_as_markdown, reset::negation::ResetNegation};

fn main() -> Result<(), RHDLError> {
    let input = (0..15)
        .map(|_| rand::random::<bool>())
        .map(|b| signal(reset_n(b)))
        .without_reset()
        .clock_pos_edge(100)
        .map(|t| t.map(|x| x.1));
    let uut = ResetNegation::<Red>::default();
    let vcd = uut.run(input)?.collect::<Vcd>();
    write_svg_as_markdown(vcd, "reset_negation.md", SvgOptions::default())?;
    Ok(())
}

With the trace

Structs

ResetNegation

The ResetNegation core. Both the input and output signals must belong to the same clock domain C.