rhdl_fpga::rng

Module xorshift

A Pseudorandom Number Generator based on XORSHIFT128

A port of the LFSR from Alchitry.com’s Lucid module pn_gen. The default seed is the same as the one in the Lucid module. Any non-zero seed will work. It takes a boolean signal as the “next” signal to advance the LFSR. - This in turn is actually a gateware implementation of XORSHIFT128.

Schematic symbol

The usage is simple. If input is true, then output will advance to the next pseudo-random number in the sequence. Otherwise, the output is latched by internal flip flops.

Example

This core is simple to use.

use rhdl::prelude::*;
use rhdl_fpga::{doc::write_svg_as_markdown, rng::xorshift::XorShift};

fn main() -> Result<(), RHDLError> {
    let input = std::iter::repeat_n(true, 10)
        .with_reset(1)
        .clock_pos_edge(100);
    let uut = XorShift::default();
    let vcd = uut.run(input)?.collect::<Vcd>();
    write_svg_as_markdown(
        vcd,
        "xor_png.md",
        SvgOptions::default().with_filter("(^top.input(.*))|(^top.outputs(.*))|(^top.reset)"),
    )?;
    Ok(())
}

Here is the trace.

Structs

XorShift

The XorShift core. Note that resetting the core resets the sequence.

XorShift128

For testing, it’s handy to have a way to generate the same sequence as the hardware. This struct impl Iterator and yields the same sequence as the hardware with the default seed.