rhdl_fpga

Module stream

Stream Cores

Stream cores are used to implement stream processing, in which data elements flow and are transformed in the design. They are composable, like iterators, and are meant to run in high performance designs (and so carefully implement the principles of latency insensitive designs). The stream cores all have registered inputs and outputs, so as to avoid combinatorial pathways between the input and output. Furthermore the stream cores implement backpressure, providing both a ready signal to upstream cores, and accepting a ready signal from downstream cores. Finally, the input and output can be voided using Option. The protocol implemented in the handshake is identical to the Ready/Valid protocol from the AXI spec for channels.

Modules

chunked

Chunked Stream Core

fifo_to_stream

A FIFO-to-Stream buffer

filter

Filter Stream Core

filter_map

Filter Map Stream Core

flatten

Flatten Core

map

Map Stream Core

pipe_wrapper

Wrap a Pipe into a Stream

stream_buffer

Carloni Buffer with Option Interface

stream_to_fifo

A Stream-to-FIFO buffer

tee

Tee Stream Core

testing

Cores useful for testing streams.

xfer

Xfer Detector Core

zip

Zip Stream Core

Structs

Ready

A typed ready signal for a stream of type T.

StreamIO

A generic Stream IO type that holds a data and ready signal. Note that in a Stream interface, these signals are generally in opposite directions. So a typical Stream core will look like this:

Functions

ready

Helper function to convert a raw signal into a typed Ready

ready_cast

Helper function to cast from one ready signal to another