rhdl_fpga::stream

Module fifo_to_stream

A FIFO-to-Stream buffer

Purpose

A FIFO-to-Stream buffer is a highly specialized two element FIFO backed with a pair of registers instead of a BRAM. The idea is to allow logic that has “push” semantics like a FIFO interface, to connecto to stream, which has “pull” semantics. Both interfaces support backpressure (the FIFO via the full signal, and the stream via the ready signal).

The other way to conceptualize this is as a source and sink pair. The supply side pipeline is a data source - it produces data elements at it’s own pace. The demand side pipeline is a data sink - it consumes data elements at it’s own pace. The push pull buffer in general would be a FIFO, and indeed the Carloni papers show FIFOs as the implementation of push-pull buffers. However, in many cases, we only need minimal buffering, and a pair of registers is sufficient.

Note that a normal synchronous FIFO as included in rhdl will not work here - if it has only two slots in it, it cannot (by design) fill both slots.

The design of the buffer uses a state to manage the fill level, and uses the extra value of a fill level of 3 to indicate that the push-pull buffer is in an error condition due to overflow of the input. To make this buffer easy to use with Carloni skid buffers, the output is presented as an Option<T>, and underflow is not possible, as None is returned when the buffer is empty.

Note that one use case for the FIFOToStream buffer is when we need to be able to anticipate by a clock cycle that a pipeline is able to push data forward. Here is an example of the problem:

During the time t1, the downstream pipeline was available for us to push a new data item, but our upstream process was not ready. In time t2, the downstream pipeline is no longer available, but the upstream process has produced a data item. The upstream process must stall and hold this output value until the downstream pipeline re-raises ready for a clock cycle.

With the FIFOToStream buffer, we have an addition invariant:

A FIFO that is not full on cycle T cannot be full on cycle T+1 if we do not add data to it.

This invariant means that the equivalent timing diagram with a FIFOToStream buffer looks like this instead

The important difference is that if the input stage is !full, as happens in interval t1, the upstream pipeline is guaranteed to be able to run and produce a data item, even if it is in the future. Thus, if the upstream pipeline waits for the output to be !full, it can gaurantee that one output item can be produced.

Schematic Symbol

Here is the schematic symbol for the FIFOToStream buffer

Internals

Effectively, the FIFOToStream buffer is simply a 2-element FIFO. It is implemented with a pair of registers and manual control logic, since the general FIFO logic does not handle such small sizes well.

Roughly the internal circuitry is equivalent to this:

The FIFO is advanced only if the output is `Some`, and if the `ready` signal is asserted.
Note that there are no combinatorial paths between the inputs and outputs, and a test is used to verify this property.

Example

Here is an example of the interface.

use rhdl::prelude::*;
use rhdl_fpga::{
    rng::xorshift::XorShift128,
    stream::fifo_to_stream::{self, FIFOToStream},
};

fn main() -> Result<(), RHDLError> {
    // The buffer will manage items of 4 bits
    let uut = FIFOToStream::<b4>::default();
    // The test harness will include a consumer that
    // randomly pauses the upstream producer.
    let mut need_reset = true;
    let mut source_rng = XorShift128::default();
    let vcd = uut
        .run_fn(
            |out| {
                if need_reset {
                    need_reset = false;
                    return Some(rhdl::core::sim::ResetOrData::Reset);
                }
                let mut input = fifo_to_stream::In::<b4>::dont_care();
                let want_to_pause = rand::random::<u8>() > 200;
                input.ready.raw = !want_to_pause;
                // Decide if the producer will generate a data item
                let want_to_send = rand::random::<u8>() < 200;
                input.data = None;
                if !out.full && want_to_send {
                    input.data = source_rng.next().map(|x| bits((x & 0xF) as u128));
                }
                Some(rhdl::core::sim::ResetOrData::Data(input))
            },
            100,
        )
        .take_while(|t| t.time < 1500)
        .collect::<Vcd>();
    rhdl_fpga::doc::write_svg_as_markdown(
        vcd,
        "fifo_to_rv.md",
        SvgOptions::default().with_io_filter(),
    )?;
    Ok(())
}

With an output.

Structs

FIFOToStream

The FIFOToStream Buffer core.

Out

Outputs from the FIFOToStream buffer

Type Aliases

In

Inputs to the FIFOToStream buffer