rhdl_fpga::stream

Module filter

Filter Stream Core

Purpose

A Filter Core takes a stream of elements of type T and a function fn(T) -> bool, and keeps only those items for which the function evaluates to true. The filter function is provided in the form of a synthesizable function. This is equivalent to using .filter() on an interator.

Schematic Symbol

Here is the schematic symbol for the Filter core

Internals

Unlike [Flatten] or [Chunked], the [FilterPipe] does not impose any flow control on the upstream. Because it can at most produce as many items as the source, it can be implemented with a StreamBuffer buffers at the input which is needed to isolate the combinatorial filter function from the remaining parts of the stream.
Note that if you need a more expensive filter function (i.e., one that itself is pipelined), then you cannot use this construct.

# Example
Here is an example of filtering a stream.

use std::iter::repeat_n;

use rhdl::prelude::*;
use rhdl_fpga::{
    rng::xorshift::XorShift128,
    stream::{
        filter::Filter,
        testing::{single_stage::single_stage, utils::stalling},
    },
};

#[kernel]
fn keep_even(_cr: ClockReset, t: b4) -> bool {
    !(t & bits(1)).any()
}

fn main() -> Result<(), RHDLError> {
    let a_rng = XorShift128::default().map(|x| b4((x & 0xF) as u128));
    let a_rng = stalling(a_rng, 0.23);
    let consume = move |data: Option<b4>| {
        if let Some(data) = data {
            // Only even values kept
            assert!(data.raw() & 1 == 0);
        }
        rand::random::<f64>() > 0.2
    };
    let filter = Filter::try_new::<keep_even>()?;
    let uut = single_stage(filter, a_rng, consume);
    // Run a few samples through
    let input = repeat_n((), 15).with_reset(1).clock_pos_edge(100);
    let vcd = uut.run_without_synthesis(input)?.collect::<Vcd>();
    rhdl_fpga::doc::write_svg_as_markdown(vcd, "filter.md", SvgOptions::default())?;
    Ok(())
}

with a trace file like this:

Structs

Filter

The Filter Stream Core

Type Aliases

In

The input for the Filter

Out

The output of the Filter