rhdl_fpga::stream::filter

Struct Filter

pub struct Filter<T: Digital> { /* private fields */ }

The Filter Stream Core

Here T is the type flowing in the stream. At construction time, you provide a synthesizable function to filter the contents of the stream. Only items for which fn(T) returns true will be passed on downstream.

Implementations

impl<T> Filter<T>
where T: Digital,

pub fn try_new<S>() -> Result<Self, RHDLError>
where S: DigitalFn + DigitalFn2<A0 = ClockReset, A1 = T, O = bool>,

Construct a Filter Stream

The argument to the filter try_new function is a synthesizable function (i.e., one marked with the #[kernel] attribute). It must have a signature of fn(ClockReset, T) -> bool.