rhdl_fpga::stream

Module pipe_wrapper

Wrap a Pipe into a Stream

Purpose

This core is used to take a pipeline with no backpressure, and interface it into a stream. The backpressure is handled by an internal FIFO where output elements of the pipe are allocated space (like a credit-based system in networking).

Details

The original Latency Insensitive Design work focused on stallable pipelines. That is to say, that if the ready signal was taken away (or equivalently, if the downstream process asserted stop), then the entire pipeline would stall until the ready signal was reasserted. In the original papers, this was done via a gated clock or a clock-enable signal that was used to either advance a given stage in the pipeline or hold it in it’s current state. Roughly, something like this:

However, the idea of a clk_en line doesn’t always fit with a pipeline. For example, a DRAM controller can be seen as a pipeline (where S is the address to read from and T are the data elements read back, for example). The DRAM controller is generally not stallable. And it is fair to assume that the controller requires you to read out the data elements from the output once you have committed a certain transaction.

Furthermore, suppose that each item S injected into the pipeline produces N items of type T on the output of the pipeline. Then when can a new item be injected? If the pipeline is opaque, then we can only keep track of how many pending items need to be written to the output.

The obvious answer is to include an output FIFO at the end of the pipeline to hold the items as they are produced by the pipeline. These can then be served to the downstream process as it manages the ready signal.

Ignoring, temporarily the problem of underflow of the output FIFO, the bigger problem is the lack of backpressure handling by the pipeline. If the output FIFO is full, how do we stall the pipeline? If it has no clock enable or other means of stalling, we are still in the same situation as before.

The proposed solution in this core is to introduce a credit-based system. A control core that tracks the number of open slots in the output FIFO, and only dispatches as many items S such that the output is guaranteed to fit in the output FIFO. Each clock for which the ready signal is asserted will release an additional credit to the controller, and each S item that is consumed will require N credits to be available, where N is the number of T items produced by each S.

Thus, backpressure is moved upstream of the pipeline. The pipeline itself does not need to support backpressure, since the controller will stop the inflow of data when there is insufficient credit in the FIFO to start processing more data elements.

Furthermore this design is invariant to the latency introduced by the pipeline. It can even be variable. Each output slot in the FIFO is reserved for a pending computation, and credit tracking makes no assumptions about how long those reservations are held for.

Schematic Symbol

Here is the schematic symbol for the PipeWrapper.

It is understood that the pipline will start when fed Some(S) data element, and will produce exactly one Option output element at some time in the future. The internal FIFO size is exposed, since knowledge of how big the output FIFO will need to be is a design decision.

Example

An example of wrapping a pipeline with a PipeWrapper core is here.

use std::iter::repeat_n;

use badascii_doc::badascii;
use rhdl::prelude::*;

use delay::DelayLine;
use rhdl_fpga::{
    core::slice::lsbs,
    rng::xorshift::XorShift128,
    stream::{
        pipe_wrapper::PipeWrapper,
        testing::{sink_from_fn::SinkFromFn, source_from_fn::SourceFromFn, utils::stalling},
    },
};

pub mod delay {
    use rhdl_fpga::core::{
        dff::DFF,
        option::{pack, unpack},
        slice::lsbs,
    };

    use super::*;
    #[derive(Clone, Synchronous, SynchronousDQ, Default)]
    pub struct DelayLine {
        stage_0: DFF<Option<b6>>,
        stage_1: DFF<Option<b6>>,
        stage_2: DFF<Option<b4>>,
    }

    impl SynchronousIO for DelayLine {
        type I = Option<b6>;
        type O = Option<b4>;
        type Kernel = kernel;
    }

    #[kernel]
    pub fn kernel(_cr: ClockReset, i: Option<b6>, q: Q) -> (Option<b4>, D) {
        let mut d = D::dont_care();
        d.stage_0 = i;
        d.stage_1 = q.stage_0;
        let (tag, data) = unpack::<b6>(q.stage_1, bits(0));
        let data = lsbs::<U4, U6>(data);
        d.stage_2 = pack::<b4>(tag, data);
        (q.stage_2, d)
    }
}

///
/// Here is a sketch of the internals:
///
#[doc = badascii!(r"
+Source+-+    +Wrapper+-----+     +Sink+--+
|        | ?T |             | ?S  |       |
|    data+--->|data     data+---->|data   |
|        |    |             |     |       |
|   ready|<---+ready   ready|<----+ready  |
+--------+    +--+------+---+     +-------+
           +-----+      +----+             
         ?T|  +------------+ |?S           
           +->|in       out+-+             
              +------------+               
")]
#[derive(Clone, Synchronous, SynchronousDQ)]
struct TestFixture {
    source: SourceFromFn<b6>,
    delay: DelayLine,
    wrapper: PipeWrapper<b6, b4, U2>,
    sink: SinkFromFn<b4>,
}

impl SynchronousIO for TestFixture {
    type I = ();
    type O = ();
    type Kernel = kernel;
}

#[kernel]
#[doc(hidden)]
pub fn kernel(_cr: ClockReset, _i: (), q: Q) -> ((), D) {
    let mut d = D::dont_care();
    d.wrapper.data = q.source;
    d.source = q.wrapper.ready;
    d.sink = q.wrapper.data;
    d.wrapper.ready = q.sink;
    d.delay = q.wrapper.to_pipe;
    d.wrapper.from_pipe = q.delay;
    ((), d)
}

fn main() -> Result<(), RHDLError> {
    let b_rng = XorShift128::default().map(|x| b6(((x >> 8) & 0x3F) as u128));
    let mut c_rng = b_rng.clone();
    let b_rng = stalling(b_rng, 0.13);
    let consume = move |data| {
        if let Some(data) = data {
            let validation = lsbs::<U4, U6>(c_rng.next().unwrap());
            assert_eq!(data, validation);
        }
        rand::random::<f64>() > 0.2
    };
    let uut = TestFixture {
        source: SourceFromFn::new(b_rng),
        delay: DelayLine::default(),
        wrapper: PipeWrapper::default(),
        sink: SinkFromFn::new(consume),
    };
    // Run a few samples through
    let input = repeat_n((), 15).with_reset(1).clock_pos_edge(100);
    let vcd = uut.run_without_synthesis(input)?.collect::<Vcd>();
    rhdl_fpga::doc::write_svg_as_markdown(
        vcd,
        "pipe_wrap.md",
        SvgOptions::default().with_filter("(.*wrapper.input.data)|(.*wrapper.outputs.data)"),
    )?;
    Ok(())
}

With the resulting trace.

Structs

In

Inputs for the PipeWrapper

Out

Outputs from the PipeWrapper

PipeWrapper

PipeWrapper core for wrapping a pipeline into a stream