rhdl_fpga::stream

Module stream_buffer

Carloni Buffer with Option Interface

This core wraps the Carloni skid buffer core in a more ergonomic Option based interface. The void_in input and the data_in are combined into a single data_in with type Option, and the void_out and data_out are similarly combined into a single data_out with type Option. Furthermore for compatibility with the ready-valid interface used elsewhere in RHDL, the stop signals are inverted to be valid signals.

Schematic symbol

Here is the symbol for the buffer.

Internal details

Internally, the buffer is simply a [CarloniBuffer] with pack and unpack cores to convert the Option to a pair of data and valid lines. The code is pretty short and self-expanatory.

Here is a sketch of the internals

Example

Here is the example from the [CarloniBuffer] with an Option based interface.

use rhdl::prelude::*;
use rhdl_fpga::{
    rng::xorshift::XorShift128,
    stream::stream_buffer::{self, StreamBuffer},
};

fn main() -> Result<(), RHDLError> {
    // The buffer will manage items of 4 bits
    let uut = StreamBuffer::<b4>::default();
    // Create a random data stream that pauses randomly
    let mut source_rng = XorShift128::default();
    let mut send = (0..).map(move |_| {
        if rand::random::<u8>() < 200 {
            source_rng.next().map(|x| bits((x & 0xF) as u128))
        } else {
            None
        }
    });
    let mut need_reset = true;
    let vcd = uut
        .run_fn(
            |out| {
                if need_reset {
                    need_reset = false;
                    return Some(rhdl::core::sim::ResetOrData::Reset);
                }
                let mut input = stream_buffer::In::<b4>::dont_care();
                // if the valid flag is high, then advance the source RNG (maybe)
                input.data = None;
                if out.ready.raw {
                    input.data = send.next().unwrap();
                }
                input.ready.raw = rand::random::<u8>() < 170;
                Some(rhdl::core::sim::ResetOrData::Data(input))
            },
            100,
        )
        .take_while(|t| t.time < 1500)
        .collect::<Vcd>();
    rhdl_fpga::doc::write_svg_as_markdown(
        vcd,
        "option_carloni.md",
        SvgOptions::default().with_io_filter(),
    )?;
    Ok(())
}

With trace

Structs

StreamBuffer

Option-based Carloni buffer core

Type Aliases

In

Inputs to the StreamBuffer buffer core

Out

Outputs from the StreamBuffer buffer core