rhdl_fpga::stream::stream_buffer

Type Alias In

pub type In<T> = StreamIO<T, T>;

Inputs to the StreamBuffer buffer core

Aliased Type

pub struct In<T> {
    pub data: Option<T>,
    pub ready: Ready<T>,
}

Fields
data: Option<T>

The data either flowing into or out of the Pipe core

ready: Ready<T>

The ready signal either flowing into or out of the Pipe core