rhdl_fpga::stream

Module stream_to_fifo

A Stream-to-FIFO buffer

Purpose

A Stream-to-FIFO buffer is a highly specialized two element FIFO backed with a pair of registers instead of a BRAM.
Note that a FIFO cannot be interfaced to a stream by simply setting ready = !full. This is because the FIFO interface contract says that setting data = Some(_) when full = true leads to an overflow. The particular use of this StreamToFIFO buffer is to minimize the number of resources needed. As it only requires a couple of registers, it is generally far less resource intensive than a full FIFO.

Schematic Symbol

Here is the schematic symbol for the StreamToFIFO buffer

Internals

Effectively, the StreamToFIFO buffer is simply a 2-element FIFO. It is implemented with a pair of registers and manual control logic, since the general FIFO logic doesn’t work with such small sizes.

Roughly the internal circuitry is equivalent to this:

Unlike the general FIFOs, it will not overflow, as it assumes that the data signal will be held until the `Ready` signal is provided. This is different from normal FIFOs, which will overflow under those conditions. The FIFO will signal that it is `ready` as long as it is not `full`. The consumer can use the `next` signal to accept the current `data` element.
Note that there are no combinatorial paths between the inputs and outputs, and a test is used to verify this property.

Example

Here is an example of the buffer in action.

use rhdl::prelude::*;
use rhdl_fpga::{rng::xorshift::XorShift128, stream::stream_to_fifo::StreamToFIFO};

fn main() -> Result<(), RHDLError> {
    // The buffer will manage items of 4 bits
    let uut = StreamToFIFO::<b4>::default();
    // The test harness will include a consumer that
    // randomly pauses the upstream producer.
    let mut need_reset = true;
    let mut source_rng = XorShift128::default().map(|x| bits((x & 0xF) as u128));
    let mut dest_rng = source_rng.clone();
    let mut source_datum = source_rng.next();
    let vcd = uut
        .run_fn(
            |out| {
                if need_reset {
                    need_reset = false;
                    return Some(rhdl::core::sim::ResetOrData::Reset);
                }
                let mut input = rhdl_fpga::stream::stream_to_fifo::In::<b4>::dont_care();
                let may_accept = rand::random::<u8>() > 150;
                let will_accept = may_accept & out.data.is_some();
                input.next = false;
                if will_accept {
                    assert_eq!(out.data, dest_rng.next());
                    input.next = true;
                }
                let will_offer = rand::random::<u8>() > 150;
                if will_offer {
                    input.data = source_datum;
                } else {
                    input.data = None;
                }
                let will_advance = will_offer & out.ready.raw;
                if will_advance {
                    source_datum = source_rng.next();
                }
                Some(rhdl::core::sim::ResetOrData::Data(input))
            },
            100,
        )
        .take_while(|t| t.time < 1500)
        .collect::<Vcd>();
    rhdl_fpga::doc::write_svg_as_markdown(
        vcd,
        "rv_to_fifo.md",
        SvgOptions::default().with_io_filter(),
    )?;
    Ok(())
}

With the output.

Structs

In

Inputs to the StreamToFIFO buffer

Out

Outputs from the StreamToFIFO buffer

StreamToFIFO

The StreamToFIFO Buffer Core