rhdl_fpga::stream

Struct StreamIO

pub struct StreamIO<T: Digital, S: Digital> {
    pub data: Option<T>,
    pub ready: Ready<S>,
}

A generic Stream IO type that holds a data and ready signal. Note that in a Stream interface, these signals are generally in opposite directions. So a typical Stream core will look like this:

In this case, we would have that

 <UUT as SynchronousIO>::In == StreamIO<T, S>
 <UUT as SynchronousIO>::Out == StreamIO<S, T>

This type exists so that cores can be reused by constraining the input and output types.

Fields
data: Option<T>

The data either flowing into or out of the Pipe core

ready: Ready<S>

The ready signal either flowing into or out of the Pipe core