rhdl_fpga::stream::testing

Module lazy_random

A Lazy Random Stream Source

This source stream generates a sequence of 32-bit pseudo-random numbers with a probability of “sleeping” after generating each one. This mimics a stalling source. Note that this is completely synthesizable.

Schematic Symbol

The lazy stream is very easy to use:

If you need something other than a [b32] output, then use the [crate::stream::map::Map] to map the stream to a different type.
Internal Details

The LazyRng core is simply a FIFOFiller coupled to a FIFOToStream core that implements the stream interface.

Roughly:

The probability of sleeping, and the duration of the sleep are controlled at construction time and are passed to the constructor of the [FIFOFiller] core. # Example
Here is an example of generating a bursty stream of random values.

use rhdl::prelude::*;
use rhdl_fpga::stream::{ready, testing::lazy_random::*};

fn main() -> Result<(), RHDLError> {
    let input = (0..)
        .map(|_| rand::random_bool(0.8))
        .map(|r| In { ready: ready(r) })
        .with_reset(1)
        .clock_pos_edge(100)
        .take_while(|t| t.time < 1500);
    let uut = LazyRng::default();
    let vcd = uut.run(input)?.collect::<Vcd>();
    rhdl_fpga::doc::write_svg_as_markdown(
        vcd,
        "lazy_rng.md",
        SvgOptions::default().with_io_filter(),
    )?;
    Ok(())
}

with a trace file like this:

Structs

In

Input for the LazyRng the ready signal provides backpressure.

LazyRng

Lazy, bursty random number generator as a stream.

Out

Data from the LazyRng stream. Will be None if the source is stalling.