rhdl_fpga::stream::testing

Module single_stage

Single Stage Test Fixture

In this test fixture, we assemble a source, a uut, and a sink into a single assembly:

The source is provided by an iterator that yields [Option] values, which are injected into the UUT when `ready` is `true`. The sink is provided by a consumer function that is called with each accepted data element as a [Option], with the `Some` variant indicating that the data element is valid. The consumer function provides the `ready` signal back to the stream to supply backpressure.

Structs

SingleStage

The SingleStage test fixture

Functions

single_stage

Create a single stage test fixture