rhdl_fpga::stream::testing::single_stage

Function single_stage

pub fn single_stage<S, T, C>(
    uut: C,
    source: impl Iterator<Item = Option<S>> + 'static,
    sink: impl FnMut(Option<T>) -> bool + 'static,
) -> SingleStage<S, T, C>
where
    S: Digital,
    T: Digital,
    C: Synchronous<I = StreamIO<S, T>, O = StreamIO<T, S>>,

Create a single stage test fixture

uut is the stream stage to be tested source is an iterator that returns items of type Option<S> sink is a function that consumes elements of type Option<T> and provides the backpressure signal as a return.

To see example usages, look at the [Flatten] and [Chunked] examples.