rhdl_fpga::stream::testing

Module sink_from_fn

Test Sink From Function

Purpose

For testing stream processes, it’s often handy to have a sink for a stream that can be generated from a closure without worrying that something that is synthesizable.

Structs

SinkFromFn

The SinkFromFn core