rhdl_fpga::stream::testing::utils

Function stalling

pub fn stalling<S>(
    s: S,
    prob: f64,
) -> impl Iterator<Item = Option<<S as Iterator>::Item>>
where
    S: Iterator,

The stalling function wraps an iterator into one that “stalls” at random, returning either Some(t) (where t is the value yielded by the underlying iterator) or None. The probability of a stall is a parameter prob.