rhdl_fpga::stream

Module xfer

Xfer Detector Core

Purpose

A Xfer Core sits on a stream, and emits a pulse (combinatorially) in each clock cycle that a valid transfer takes place on the stream. It is equivalent to computing data.is_some() && ready, but can be easier to use as a block than as an expression.

Schematic Symbol

Here is the schematic symbol for the Xfer core

Internals

This core is very simple. It passes the data and ready signals through, and derives (combinatorially) the run signal as data.is_some() && ready.

Example

Structs

Out

Output of the Xfer core

Xfer

The Xfer core